+++ src/main/java/dev/anvilcraft/resource/ageratum/client/feat/markdown/component/MDNoticeBoxComponent.java
@@ -52,7 +52,7 @@ public class MDNoticeBoxComponent extends MDComponent {
     }
 
     @Override
-    public void render(GuiGraphics guiGraphics, Minecraft minecraft, int maxX, int maxY, int mouseX, int mouseY) {
+    public void render(GuiGraphics guiGraphics, Minecraft minecraft, int maxX, int maxY, float mouseX, float mouseY) {
         if (this.contentComponents.isEmpty()) {
             return;
         }
